ai-recruiter/pyproject.toml
[project]
name = "ai-recruiter"
version = "0.2.0"
description = "Unified AI-powered recruitment platform: pre-screening (matching) + multimodal interview (NLP, vision/behavioral, avatar) under a single FastAPI app."
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}

dependencies = [
    # ── Core ────────────────────────────────────────────────────────────────
    "pydantic>=2.6",
    "httpx>=0.25",
    "redis>=5.0",
    "python-dotenv>=1.0",
    "pyyaml>=6.0",
    "structlog>=23.2",
    "numpy<2",

    # ── Agent framework ─────────────────────────────────────────────────────
    "langgraph>=0.0.40",
    "langchain>=0.1",

    # ── Serving ─────────────────────────────────────────────────────────────
    "fastapi>=0.110",
    "uvicorn[standard]>=0.27",
    "python-multipart>=0.0.9",
    "websockets>=12.0",

    # ── ML common ───────────────────────────────────────────────────────────
    "torch>=2.1",
    "transformers>=4.36",
    "sentence-transformers>=2.2",
    "datasets>=2.14",
    "accelerate>=0.29",
    "peft>=0.7",
    "scikit-learn>=1.3",
    "scipy",
    "joblib>=1.3",
    "tqdm>=4.65",

    # ── Matching agent (Paper) ──────────────────────────────────────────────
    "torch-geometric>=2.4",
    "pypdf>=4.0",
    "python-docx>=1.1",
    "pymupdf>=1.24",

    # ── Vision agent (stress detection) ─────────────────────────────────────
    "opencv-python>=4.8",
    "mediapipe>=0.10",
    "hsemotion>=0.3.0",
    "timm>=0.9,<1.0",
    "reportlab>=4.0",
    "matplotlib>=3.7",

    # ── NLP agent ───────────────────────────────────────────────────────────
    # vllm is intentionally NOT listed: install separately with `pip install vllm`
    # so its CUDA-tied resolver doesn't slow this file.
    "kokoro",
    "soundfile",
    "faster-whisper",
    "sentencepiece",
    "protobuf",

    # ── Communication / extras ──────────────────────────────────────────────
    "grpcio>=1.60",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4",
    "pytest-asyncio>=0.23",
    "pytest-cov>=4.1",
    "ruff>=0.1",
    "mypy>=1.8",
    "httpx>=0.25",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

# Hatchling can't auto-detect packages because the distribution name
# (`ai-recruiter`) doesn't match a top-level directory. List them explicitly.
[tool.hatch.build.targets.wheel]
packages = ["core", "agents", "configs"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
